requirements.txt
appdirs==1.4.3
attrs==19.3.0
autopep8==1.5
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-WTF==0.14.3
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
WTForms==2.2.1
zeep==3.4.0
python-dotenv==0.12.0
